options/linux: moved poll.h to sys/poll.h
